src/requirements.txt
aiohttp==3.8.6
discord.py==2.3.2
numpy==1.24.1
opencv_python==4.7.0.68
Pillow==10.3.0
python-dotenv==1.0.1
Requests==2.31.0
torch==2.1.0
websocket-client==1.7.0
